-release

+++ Nice Day/Extensions/Label+extension.swift
@@ -85,7 +85,7 @@ extension UILabel {
     - Parameters:
        - for: Значение активности, которое будет отображаться в Cell
     */
-    func setAttributedStringForSearch(for xpCount: Int) {
+    func setAttributedStringForSearch(for xpCount: Int16) {
         let str = "\(xpCount) xp \("_every_minute".localized)"
         let attributedString = NSMutableAttributedString(string: str, attributes: [
             .font: UIFont.systemFont(ofSize: 12.0, weight: .semibold),
